requirements.txt
requests
PyYAML
numpy
chromadb
ollama
sentence-transformers
nltk
scikit-learn
pandas
python-dotenv